Reapply "Changes made"
This reverts commit 06b20fca9413ff6a828d5b2f14397ce2ac37bcf3.
