+++ preliz/internal/optimization.py
@@ -451,8 +451,37 @@ def find_kappa(data, mu):
     else:
         return np.finfo(float).tiny
 
+def find_mode_logitnormal(distribution):
+    def mode_equation(x):
+        # The equation is: logit(x) = σ²(2x-1) + μ
+        # We want to find the root of: logit(x) - σ²(2x-1) - μ = 0
+        return logit(x) - (distribution.sigma**2 * (2*x - 1)) - distribution.mu     
+    #Left side
+    try:
+        sol1 = root_scalar(mode_equation, bracket=(eps, 0.5-eps)).root
+    except ValueError:
+        sol1 = None
+        
+    #Right side
+    try:
+        sol2 = root_scalar(mode_equation, bracket=(0.5+eps, 1-eps)).root
+    except ValueError:
+        sol2 = None
 
-def find_mode_through_optimization(dist, bounds=None):
+    if sol1 is None and sol2 is None:
+        # If no solutions found, return the median as an approximation
+        return distribution.median()
+    elif sol1 is None:
+        return sol2
+    elif sol2 is None:
+        return sol1
+    else:
+        # Return the solution with higher density
+        if distribution.pdf(sol1) >= distribution.pdf(sol2):
+            return sol1
+        return sol2
+
+def find_mode(distribution, bounds=None):
     """Find mode of a distribution through numerical optimization.
     
     Parameters
@@ -469,10 +498,10 @@ def find_mode_through_optimization(dist, bounds=None):
         Mode of the distribution
     """
     def negative_pdf(x):
-        return -dist.pdf(x)
+        return -distribution.pdf(x)
     
     if bounds is None:
-        bounds = (0, dist.ppf(0.9999))
+        bounds = distribution._finite_endpoints("full")
         
     result = minimize_scalar(negative_pdf, bounds=bounds, method='bounded')
     return result.x
